+++ Extension/ExtensionTrait.php
@@ -11,7 +11,6 @@
 
 namespace Symfony\Component\DependencyInjection\Extension;
 
-use Symfony\Component\Config\Builder\ConfigBuilderGenerator;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\Config\Loader\DelegatingLoader;
 use Symfony\Component\Config\Loader\LoaderResolver;
@@ -51,12 +50,11 @@ trait ExtensionTrait
 
     private function createContainerLoader(ContainerBuilder $container, string $env, bool $prepend): DelegatingLoader
     {
-        $buildDir = $container->getParameter('kernel.build_dir');
         $locator = new FileLocator();
         $resolver = new LoaderResolver([
             new YamlFileLoader($container, $locator, $env, $prepend),
             new IniFileLoader($container, $locator, $env),
-            class_exists(ConfigBuilderGenerator::class) ? new PhpFileLoader($container, $locator, $env, new ConfigBuilderGenerator($buildDir), $prepend) : new PhpFileLoader($container, $locator, $env, $prepend),
+            new PhpFileLoader($container, $locator, $env, $prepend),
             new GlobFileLoader($container, $locator, $env),
             new DirectoryLoader($container, $locator, $env),
             new ClosureLoader($container, $env),