+++ Loader/PhpFileLoader.php
@@ -11,19 +11,12 @@
 
 namespace Symfony\Component\DependencyInjection\Loader;
 
-use Symfony\Component\Config\Builder\ConfigBuilderGenerator;
-use Symfony\Component\Config\Builder\ConfigBuilderGeneratorInterface;
-use Symfony\Component\Config\Builder\ConfigBuilderInterface;
-use Symfony\Component\Config\FileLocatorInterface;
 use Symfony\Component\Config\Loader\LoaderResolver;
 use Symfony\Component\DependencyInjection\Attribute\When;
 use Symfony\Component\DependencyInjection\Attribute\WhenNot;
-use Symfony\Component\DependencyInjection\Container;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Exception\InvalidArgumentException;
 use Symfony\Component\DependencyInjection\Exception\LogicException;
-use Symfony\Component\DependencyInjection\Extension\ConfigurationExtensionInterface;
-use Symfony\Component\DependencyInjection\Extension\ExtensionInterface;
 use Symfony\Component\DependencyInjection\Loader\Configurator\App;
 use Symfony\Component\DependencyInjection\Loader\Configurator\AppReference;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
@@ -39,24 +32,6 @@ use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigura
 class PhpFileLoader extends FileLoader
 {
     protected bool $autoRegisterAliasesForSinglyImplementedInterfaces = false;
-
-    private ?ConfigBuilderGeneratorInterface $generator;
-
-    public function __construct(
-        ContainerBuilder $container,
-        FileLocatorInterface $locator,
-        ?string $env = null,
-        ConfigBuilderGeneratorInterface|bool|null $generator = null,
-        bool $prepend = false,
-    ) {
-        if (\is_bool($generator)) {
-            $prepend = $generator;
-            $generator = null;
-        }
-        $this->generator = $generator;
-
-        parent::__construct($container, $locator, $env, $prepend);
-    }
 
     public function load(mixed $resource, ?string $type = null): mixed
     {
@@ -160,7 +135,6 @@ class PhpFileLoader extends FileLoader
     {
         $callback = $callback(...);
         $arguments = [];
-        $configBuilders = [];
         $r = new \ReflectionFunction($callback);
 
         $excluded = true;
@@ -217,14 +191,7 @@ class PhpFileLoader extends FileLoader
                     }
                     // no break
                 default:
-                    try {
-                        $configBuilder = $this->configBuilder($type);
-                    } catch (InvalidArgumentException|\LogicException $e) {
-                        throw new \InvalidArgumentException(\sprintf('Could not resolve argument "%s" for "%s".', $type.' $'.$parameter->getName(), $path), 0, $e);
-                    }
-                    trigger_deprecation('symfony/dependency-injection', '7.4', 'Using fluent builders for semantic configuration is deprecated, instantiate the "%s" class with the config array as argument and return it instead in "%s".', $type, $path);
-                    $configBuilders[] = $configBuilder;
-                    $arguments[] = $configBuilder;
+                    throw new \InvalidArgumentException(\sprintf('Could not resolve argument "%s" for "%s".', $type.' $'.$parameter->getName(), $path));
             }
         }
 
@@ -234,62 +201,5 @@ class PhpFileLoader extends FileLoader
         } finally {
             --$this->importing;
         }
-
-        foreach ($configBuilders as $configBuilder) {
-            $this->loadExtensionConfig($configBuilder->getExtensionAlias(), ContainerConfigurator::processValue($configBuilder->toArray()));
-        }
     }
-
-    /**
-     * @param string $namespace FQCN string for a class implementing ConfigBuilderInterface
-     */
-    private function configBuilder(string $namespace): ConfigBuilderInterface
-    {
-        if (!class_exists(ConfigBuilderGenerator::class)) {
-            throw new \LogicException('You cannot use the config builder as the Config component is not installed. Try running "composer require symfony/config".');
-        }
-
-        if (null === $this->generator) {
-            throw new \LogicException('You cannot use the ConfigBuilders without providing a class implementing ConfigBuilderGeneratorInterface.');
-        }
-
-        // If class exists and implements ConfigBuilderInterface
-        if (class_exists($namespace) && is_subclass_of($namespace, ConfigBuilderInterface::class)) {
-            return new $namespace();
-        }
-
-        // If it does not start with Symfony\Config\ we don't know how to handle this
-        if (!str_starts_with($namespace, 'Symfony\\Config\\')) {
-            throw new InvalidArgumentException(\sprintf('Could not find or generate class "%s".', $namespace));
-        }
-
-        // Try to get the extension alias
-        $alias = Container::underscore(substr($namespace, 15, -6));
-
-        if (str_contains($alias, '\\')) {
-            throw new InvalidArgumentException('You can only use "root" ConfigBuilders from "Symfony\\Config\\" namespace. Nested classes like "Symfony\\Config\\Framework\\CacheConfig" cannot be used.');
-        }
-
-        if (!$this->container->hasExtension($alias)) {
-            $extensions = array_filter(array_map(fn (ExtensionInterface $ext) => $ext->getAlias(), $this->container->getExtensions()));
-            throw new InvalidArgumentException(UndefinedExtensionHandler::getErrorMessage($namespace, null, $alias, $extensions));
-        }
-
-        $extension = $this->container->getExtension($alias);
-        if (!$extension instanceof ConfigurationExtensionInterface) {
-            throw new \LogicException(\sprintf('You cannot use the config builder for "%s" because the extension does not implement "%s".', $namespace, ConfigurationExtensionInterface::class));
-        }
-
-        $configuration = $extension->getConfiguration([], $this->container);
-        $loader = $this->generator->build($configuration);
-
-        return $loader();
-    }
-}
-
-/**
- * @internal
- */
-final class ProtectedPhpFileLoader extends PhpFileLoader
-{
 }